Implement the API needed to run evals on the web
